refactor: 캐러셀 훅 분리
