+++ src/pages/lanternDetail/constants/carousel.ts
@@ -1,0 +1,4 @@
+export const COPIES = 3;
+export const MIDDLE_BLOCK = Math.floor(COPIES / 2);
+export const COOLDOWN_MS = 500;
+export const DELTA_TRIGGER = 20;
